Move Routing to app nad create folder components.
