+++ src/components/emptyComponent/EmptyComponent2.js
@@ -1,0 +1,11 @@
+function EmptyComponent () {
+    return (
+        <div>
+            <div className="content-container">
+                <span>Work in progress</span>
+            </div>
+        </div>
+    );
+}
+
+export default EmptyComponent;
