+++ src/components/home/Home.js
@@ -1,14 +1,10 @@
 import './App.css';
-import Header from '../common/Header'
-import Footer from '../common/Footer';
 
 import Banner from './Banner';
 
-function App() {
-  return (
-    <div>
-        <Header />
-        <div className={"content-container"}>
+function Home() {
+    return (
+        <div>
             <Banner
                 message={"Advertisement"}
                 submessage={"Design + Print + Install + Delivery"}
@@ -16,9 +12,7 @@ function App() {
             />
             <hr/>
         </div>
-        <Footer />
-    </div>
-  );
+    );
 }
 
-export default App;
+export default Home;
